Extract the utf case folding code to make it accessible from Rust
