+++ src/normalize.h
@@ -1,0 +1,5 @@
+#pragma once
+
+/* Normalize sorting string for storage. This folds everything to unicode equivalent strings. The
+ * allocated return string needs to be freed later */
+char *normalizeStr(const char *str);
