+++ src/redisearch_rs/ffi/build.rs
@@ -52,6 +52,7 @@ fn main() {
         root.join("src").join("redisearch.h"),
         root.join("src").join("buffer.h"),
         root.join("src").join("result_processor.h"),
+        root.join("src").join("normalize.h"),
     ];
 
     let mut bindings = bindgen::Builder::default();
