+++ src/sortable.c
@@ -9,12 +9,12 @@
 #include <stdlib.h>
 #include <stdio.h>
 #include "rmutil/rm_assert.h"
-#include "libnu/libnu.h"
 #include "rmutil/util.h"
 #include "rmutil/strings.h"
 #include "rmalloc.h"
 #include "sortable.h"
 #include "buffer.h"
+#include "normalize.h"
 
 /* Create a sorting vector of a given length for a document */
 RSSortingVector *NewSortingVector(size_t len) {
@@ -28,37 +28,6 @@ RSSortingVector *NewSortingVector(size_t len) {
     ret->values[i] = RS_NullVal();
   }
   return ret;
-}
-
-/* Normalize sorting string for storage. This folds everything to unicode equivalent strings. The
- * allocated return string needs to be freed later */
-char *normalizeStr(const char *str) {
-
-  size_t buflen = 2 * strlen(str) + 1;
-  char *lower_buffer = rm_calloc(buflen, 1);
-  char *lower = lower_buffer;
-  char *end = lower + buflen;
-
-  const char *p = str;
-  size_t off = 0;
-  while (*p != 0 && lower < end) {
-    uint32_t in = 0;
-    p = nu_utf8_read(p, &in);
-    const char *lo = nu_tofold(in);
-
-    if (lo != 0) {
-      uint32_t u = 0;
-      do {
-        lo = nu_casemap_read(lo, &u);
-        if (u == 0) break;
-        lower = nu_utf8_write(u, lower);
-      } while (u != 0 && lower < end);
-    } else {
-      lower = nu_utf8_write(in, lower);
-    }
-  }
-
-  return lower_buffer;
 }
 
 #define RSPUT_SANITY_CHECK \
